+++ 3o14/package.json
@@ -26,6 +26,7 @@
     "date-fns": "^4.1.0",
     "eas-cli": "^14.2.0",
     "expo": "~52.0.23",
+    "expo-av": "^15.0.1",
     "expo-blur": "~14.0.1",
     "expo-constants": "~17.0.3",
     "expo-font": "~13.0.2",
